+++ messages/testmsgs/generator.php
@@ -1,0 +1,14 @@
+<?php defined('SYSPATH') OR die('No direct script access.');
+
+return array
+(
+	'one' => 'first message',
+	'two' => array
+	(
+		'three' => 'second message',
+		'four' => array
+		(
+			'five' => 'third message',
+		),
+	),
+);